refactoring help overlay page and increasing tests coverage +5%
